@@ -1,6 +1,46 @@
 {
   "name": "@web3-onboard/core",
-  "version": "2.2.6",
+  "version": "2.8.4-alpha.1",
+  "description": "Web3-Onboard makes it simple to connect Ethereum hardware and software wallets to your dapp. Features standardized spec compliant web3 providers for all supported wallets, framework agnostic modern javascript UI with code splitting, CSS customization, multi-chain and multi-account support, reactive wallet state subscriptions and real-time transaction state change notifications.",
+  "keywords": [
+    "Ethereum",
+    "Web3",
+    "EVM",
+    "dapp",
+    "Multichain",
+    "Wallet",
+    "Transaction",
+    "Provider",
+    "Hardware Wallet",
+    "Notifications",
+    "React",
+    "Svelte",
+    "Vue",
+    "Next",
+    "Nuxt",
+    "MetaMask",
+    "Coinbase",
+    "WalletConnect",
+    "Ledger",
+    "Trezor",
+    "Connect Wallet",
+    "Ethereum Hooks",
+    "Blocknative",
+    "Mempool",
+    "pending",
+    "confirmed",
+    "Injected Wallet",
+    "GameStop",
+    "Crypto",
+    "Crypto Wallet"
+  ],
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/blocknative/web3-onboard.git",
+    "directory": "packages/core"
+  },
+  "homepage": "https://www.blocknative.com/onboard",
+  "bugs": "https://github.com/blocknative/web3-onboard/issues",
   "scripts": {
     "build": "rollup -c",
     "dev": "rollup -c -w",
@@ -18,16 +58,17 @@
   "type": "module",
   "license": "MIT",
   "devDependencies": {
+    "@rollup-extras/plugin-copy": "~1.2.2",
     "@rollup/plugin-json": "^4.1.0",
     "@rollup/plugin-node-resolve": "^11.0.0",
     "@rollup/plugin-replace": "^3.0.0",
     "@rollup/plugin-typescript": "^8.0.0",
-    "@rollup-extras/plugin-copy": "~1.2.2",
     "@tsconfig/svelte": "^2.0.0",
     "@types/lodash.merge": "^4.6.6",
     "@types/lodash.partition": "^4.6.6",
     "@typescript-eslint/eslint-plugin": "^4.31.1",
     "@typescript-eslint/parser": "^4.31.1",
+    "@web3-onboard/gas": "^2.0.0",
     "eslint": "^7.32.0",
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-svelte3": "^3.2.1",
@@ -41,15 +82,18 @@
     "typescript": "^4.5.5"
   },
   "dependencies": {
-    "@web3-onboard/common": "^2.0.7",
+    "@web3-onboard/common": "^2.2.2",
+    "bignumber.js": "^9.0.0",
+    "bnc-sdk": "^4.4.1",
     "bowser": "^2.11.0",
     "ethers": "5.5.3",
     "eventemitter3": "^4.0.7",
     "joi": "17.6.0",
     "lodash.merge": "^4.6.2",
     "lodash.partition": "^4.6.0",
+    "nanoid": "^4.0.0",
     "rxjs": "^7.5.2",
-    "svelte-i18n": "^3.3.13",
-    "svelte": "^3.46.4"
+    "svelte": "^3.49.0",
+    "svelte-i18n": "^3.3.13"
   }
 }
