@@ -1,6 +1,6 @@
 {
   "name": "demo",
-  "version": "2.0.1",
+  "version": "2.0.9",
   "devDependencies": {
     "assert": "^2.0.0",
     "buffer": "^6.0.3",
@@ -14,7 +14,7 @@
     "process": "^0.11.10",
     "stream-browserify": "^3.0.0",
     "stream-http": "^3.2.0",
-    "svelte": "^3.31.2",
+    "svelte": "^3.49.0",
     "svelte-loader": "^3.0.0",
     "util": "^0.12.4",
     "webpack": "^5.64.4",
@@ -22,19 +22,24 @@
     "webpack-dev-server": "4.7.4"
   },
   "dependencies": {
-    "@web3-onboard/core": "^2.2.3",
-    "@web3-onboard/fortmatic": "^2.0.2",
-    "@web3-onboard/gnosis": "^2.0.1",
-    "@web3-onboard/injected-wallets": "^2.0.5",
-    "@web3-onboard/keepkey": "^2.0.1",
-    "@web3-onboard/keystone": "^2.0.1",
-    "@web3-onboard/ledger": "^2.0.1",
-    "@web3-onboard/portis": "^2.0.0",
-    "@web3-onboard/torus": "^2.0.1",
-    "@web3-onboard/trezor": "^2.0.2",
-    "@web3-onboard/walletconnect": "^2.0.1",
-    "@web3-onboard/coinbase": "^2.0.0",
-    "@web3-onboard/magic": "^2.0.1",
+    "@web3-onboard/coinbase": "^2.1.2",
+    "@web3-onboard/core": "^2.8.4-alpha.1",
+    "@web3-onboard/dcent": "^2.2.0",
+    "@web3-onboard/fortmatic": "^2.0.13",
+    "@web3-onboard/gas": "^2.1.2",
+    "@web3-onboard/gnosis": "^2.1.2",
+    "@web3-onboard/injected-wallets": "^2.2.2",
+    "@web3-onboard/keepkey": "^2.3.0",
+    "@web3-onboard/keystone": "^2.3.0",
+    "@web3-onboard/ledger": "^2.3.0",
+    "@web3-onboard/magic": "^2.1.2",
+    "@web3-onboard/portis": "^2.1.2",
+    "@web3-onboard/sequence": "^2.0.1",
+    "@web3-onboard/torus": "^2.1.2",
+    "@web3-onboard/trezor": "^2.3.0",
+    "@web3-onboard/tallyho": "^2.0.0",
+    "@web3-onboard/walletconnect": "^2.1.2",
+    "@web3-onboard/web3auth": "^2.1.2",
     "vconsole": "^3.9.5"
   },
   "license": "MIT",
