@@ -1,7 +1,46 @@
 {
   "name": "@web3-onboard/react",
-  "version": "2.1.5",
-  "description": "Collection of React Hooks for web3-onboard",
+  "version": "2.3.3",
+  "description": "A collection of React hooks for integrating Web3-Onboard in to React and Next.js projects. Web3-Onboard makes it simple to connect Ethereum hardware and software wallets to your dapp. Features standardised spec compliant web3 providers for all supported wallets, modern javascript UI with code splitting, CSS customization, multi-chain and multi-account support, reactive wallet state subscriptions and real-time transaction state change notifications.",
+  "keywords": [
+    "Ethereum",
+    "Web3",
+    "EVM",
+    "dapp",
+    "Multichain",
+    "Wallet",
+    "Transaction",
+    "Provider",
+    "Hardware Wallet",
+    "Notifications",
+    "React",
+    "Svelte",
+    "Vue",
+    "Next",
+    "Nuxt",
+    "MetaMask",
+    "Coinbase",
+    "WalletConnect",
+    "Ledger",
+    "Trezor",
+    "Connect Wallet",
+    "Ethereum Hooks",
+    "Blocknative",
+    "Mempool",
+    "pending",
+    "confirmed",
+    "Injected Wallet",
+    "GameStop",
+    "Crypto",
+    "Crypto Wallet"
+  ],
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/blocknative/web3-onboard.git",
+    "directory": "packages/react"
+  },
+  "homepage": "https://www.blocknative.com/onboard",
+  "bugs": "https://github.com/blocknative/web3-onboard/issues",
   "module": "dist/index.js",
   "browser": "dist/index.js",
   "main": "dist/index.js",
@@ -17,14 +56,17 @@
   },
   "license": "MIT",
   "devDependencies": {
-    "@types/react": "^17.0.39",
+    "@types/react": "^18.0.2",
+    "@types/use-sync-external-store": "^0.0.3",
+    "react": "^18.0.0",
     "typescript": "^4.5.5"
   },
   "dependencies": {
-    "@web3-onboard/common": "^2.0.7",
-    "@web3-onboard/core": "^2.2.5"
+    "@web3-onboard/core": "^2.8.4-alpha.1",
+    "@web3-onboard/common": "^2.2.2",
+    "use-sync-external-store": "1.0.0"
   },
   "peerDependencies": {
-    "react": "^17.0.2"
+    "react": ">=16.8"
   }
 }
