@@ -1,14 +1,15 @@
 import { firstValueFrom } from 'rxjs'
-import { filter, mapTo } from 'rxjs/operators'
+import { filter, map } from 'rxjs/operators'
 import { ProviderRpcErrorCode } from '@web3-onboard/common'
-import { addNewChain, switchChain } from './provider'
-import { state } from './store'
-import { switchChainModal$ } from './streams'
-import { validateSetChainOptions } from './validation'
-import type { WalletState } from './types'
+import { addNewChain, switchChain } from './provider.js'
+import { state } from './store/index.js'
+import { switchChainModal$ } from './streams.js'
+import { validateSetChainOptions } from './validation.js'
+import type { WalletState } from './types.js'
+import { toHexString } from './utils.js'
 
 async function setChain(options: {
-  chainId: string
+  chainId: string | number
   chainNamespace?: string
   wallet?: WalletState['label']
 }): Promise<boolean> {
@@ -20,10 +21,11 @@ async function setChain(options: {
 
   const { wallets, chains } = state.get()
   const { chainId, chainNamespace = 'evm', wallet: walletToSet } = options
+  const chainIdHex = toHexString(chainId)
 
   // validate that chainId has been added to chains
   const chain = chains.find(
-    ({ namespace, id }) => namespace === chainNamespace && id === chainId
+    ({ namespace, id }) => namespace === chainNamespace && id === chainIdHex
   )
 
   if (!chain) {
@@ -50,26 +52,29 @@ async function setChain(options: {
   // check if wallet is already connected to chainId
   if (
     walletConnectedChain.namespace === chainNamespace &&
-    walletConnectedChain.id === chainId
+    walletConnectedChain.id === chainIdHex
   ) {
     return true
   }
 
   try {
-    await switchChain(wallet.provider, chainId)
+    await switchChain(wallet.provider, chainIdHex)
     return true
   } catch (error) {
     const { code } = error as { code: number }
     const switchChainModalClosed$ = switchChainModal$.pipe(
       filter(x => x === null),
-      mapTo(false)
+      map(() => false)
     )
 
-    if (code === ProviderRpcErrorCode.CHAIN_NOT_ADDED) {
+    if (
+      code === ProviderRpcErrorCode.CHAIN_NOT_ADDED ||
+      code === ProviderRpcErrorCode.UNRECOGNIZED_CHAIN_ID
+    ) {
       // chain has not been added to wallet
       try {
         await addNewChain(wallet.provider, chain)
-        await switchChain(wallet.provider, chainId)
+        await switchChain(wallet.provider, chainIdHex)
         return true
       } catch (error) {
         // display notification to user to switch chain
