@@ -1,123 +1,7 @@
-import type {
-  RequestPatch,
-  AccountSelectAPI,
-  SelectAccountOptions,
-  BasePath,
-  DerivationPath,
-  Asset,
-  ScanAccounts,
-  ScanAccountsOptions,
-  AccountAddress,
-  Account,
-  AccountsList,
-  AppMetadata,
-  RecommendedInjectedWallets,
-  WalletInit,
-  WalletHelpers,
-  APIKey,
-  Device,
-  WalletModule,
-  GetInterfaceHelpers,
-  Platform,
-  DeviceOS,
-  DeviceBrowser,
-  DeviceOSName,
-  DeviceBrowserName,
-  DeviceType,
-  ChainId,
-  RpcUrl,
-  WalletInterface,
-  ProviderMessage,
-  ProviderInfo,
-  ProviderAccounts,
-  ProviderEvent,
-  SimpleEventEmitter,
-  ConnectListener,
-  DisconnectListener,
-  MessageListener,
-  ChainListener,
-  AccountsListener,
-  Balance,
-  EthAccountsRequest,
-  EthBalanceRequest,
-  EIP1102Request,
-  SelectAccountsRequest,
-  EIP3326Request,
-  EIP3085Request,
-  EthChainIdRequest,
-  EthSignTransactionRequest,
-  EthSignMessageRequest,
-  EIP712Request,
-  AddChainParams,
-  EIP1193Provider,
-  Chain,
-  TokenSymbol,
-  CustomNetwork,
-  TransactionObject
-} from './types'
+export { ProviderRpcError } from './errors.js'
+export { createEIP1193Provider } from './eip-1193.js'
+export { SofiaProLight, SofiaProRegular, SofiaProSemiBold } from './fonts.js'
+export { weiToEth } from './utils.js'
 
-export { ProviderRpcErrorCode } from './types'
-export { ProviderRpcError } from './errors'
-export { createEIP1193Provider } from './eip-1193'
-export { default as accountSelect } from './account-select'
-export { entryModal } from './entry-modal'
-export { SofiaProLight, SofiaProRegular, SofiaProSemiBold } from './fonts'
-export { getCommon } from './hdwallets'
-
-export type {
-  RequestPatch,
-  AccountSelectAPI,
-  SelectAccountOptions,
-  BasePath,
-  DerivationPath,
-  Asset,
-  ScanAccounts,
-  ScanAccountsOptions,
-  AccountAddress,
-  Account,
-  AccountsList,
-  AppMetadata,
-  RecommendedInjectedWallets,
-  WalletInit,
-  WalletHelpers,
-  APIKey,
-  Device,
-  WalletModule,
-  GetInterfaceHelpers,
-  Platform,
-  DeviceOS,
-  DeviceBrowser,
-  DeviceOSName,
-  DeviceBrowserName,
-  DeviceType,
-  ChainId,
-  RpcUrl,
-  WalletInterface,
-  ProviderMessage,
-  ProviderInfo,
-  ProviderAccounts,
-  ProviderEvent,
-  SimpleEventEmitter,
-  ConnectListener,
-  DisconnectListener,
-  MessageListener,
-  ChainListener,
-  AccountsListener,
-  Balance,
-  EthAccountsRequest,
-  EthBalanceRequest,
-  EIP1102Request,
-  SelectAccountsRequest,
-  EIP3326Request,
-  EIP3085Request,
-  EthChainIdRequest,
-  EthSignTransactionRequest,
-  EthSignMessageRequest,
-  EIP712Request,
-  AddChainParams,
-  EIP1193Provider,
-  Chain,
-  TokenSymbol,
-  CustomNetwork,
-  TransactionObject
-}
+export * from './types.js'
+export * from './validation.js'
