@@ -38,6 +38,7 @@ module.exports = {
       }
     ],
     'object-curly-spacing': ['error', 'always'],
-    '@typescript-eslint/no-empty-function': 'off'
+    '@typescript-eslint/no-empty-function': 'off',
+    quotes: ['error', 'single', { allowTemplateLiterals: true }]
   }
 }
