@@ -1,8 +1,9 @@
-import { state } from './store'
-import { removeWallet } from './store/actions'
-import { disconnectWallet$ } from './streams'
-import type { DisconnectOptions, WalletState } from './types'
-import { validateDisconnectOptions } from './validation'
+import { getBlocknativeSdk } from './services.js'
+import { state } from './store/index.js'
+import { removeWallet } from './store/actions.js'
+import { disconnectWallet$ } from './streams.js'
+import type { DisconnectOptions, WalletState } from './types.js'
+import { validateDisconnectOptions } from './validation.js'
 
 async function disconnect(options: DisconnectOptions): Promise<WalletState[]> {
   const error = validateDisconnectOptions(options)
@@ -11,6 +12,23 @@ async function disconnect(options: DisconnectOptions): Promise<WalletState[]> {
   }
 
   const { label } = options
+
+  if (state.get().notify.enabled) {
+    // handle unwatching addresses
+    const sdk = await getBlocknativeSdk()
+
+    if (sdk) {
+      const wallet = state.get().wallets.find(wallet => wallet.label === label)
+
+      wallet.accounts.forEach(({ address }) => {
+        sdk.unsubscribe({
+          id: address,
+          chainId: wallet.chains[0].id,
+          timeout: 60000
+        })
+      })
+    }
+  }
 
   disconnectWallet$.next(label)
   removeWallet(label)
