@@ -1,7 +1,7 @@
 import { fromEventPattern, Observable } from 'rxjs'
 import { filter, takeUntil, take, share, switchMap } from 'rxjs/operators'
 import partition from 'lodash.partition'
-import { utils, providers } from 'ethers'
+import { providers } from 'ethers'
 
 import type {
   ChainId,
@@ -14,16 +14,32 @@ import type {
   SelectAccountsRequest
 } from '@web3-onboard/common'
 
-import { disconnectWallet$ } from './streams'
-import type { Account, Address, Balances, Ens, WalletState } from './types'
-import { updateAccount, updateWallet } from './store/actions'
-import { validEnsChain } from './utils'
-import disconnect from './disconnect'
-import { state } from './store'
+import { weiToEth } from '@web3-onboard/common'
+import { disconnectWallet$ } from './streams.js'
+import type { Account, Address, Balances, Ens, WalletState } from './types.js'
+import { updateAccount, updateWallet } from './store/actions.js'
+import { validEnsChain } from './utils.js'
+import disconnect from './disconnect.js'
+import { state } from './store/index.js'
+import { getBlocknativeSdk } from './services.js'
 
 export const ethersProviders: {
   [key: string]: providers.StaticJsonRpcProvider
 } = {}
+
+export function getProvider(chain: Chain): providers.StaticJsonRpcProvider {
+  if (!chain) return null
+
+  if (!ethersProviders[chain.rpcUrl]) {
+    ethersProviders[chain.rpcUrl] = new providers.StaticJsonRpcProvider(
+      chain.providerConnectionInfo && chain.providerConnectionInfo.url
+        ? chain.providerConnectionInfo
+        : chain.rpcUrl
+    )
+  }
+
+  return ethersProviders[chain.rpcUrl]
+}
 
 export function requestAccounts(
   provider: EIP1193Provider
@@ -94,8 +110,8 @@ export function trackWallet(
     disconnected$
   }).pipe(share())
 
-  // when account changed, set it to first account
-  accountsChanged$.subscribe(([address]) => {
+  // when account changed, set it to first account and subscribe to events
+  accountsChanged$.subscribe(async ([address]) => {
     // no address, then no account connected, so disconnect wallet
     // this could happen if user locks wallet,
     // or if disconnects app from wallet
@@ -119,9 +135,30 @@ export function trackWallet(
         ...restAccounts
       ]
     })
+
+    // if not existing account and notifications,
+    // then subscribe to transaction events
+    if (state.get().notify.enabled && !existingAccount) {
+      const sdk = await getBlocknativeSdk()
+
+      if (sdk) {
+        const wallet = state
+          .get()
+          .wallets.find(wallet => wallet.label === label)
+        try {
+          sdk.subscribe({
+            id: address,
+            chainId: wallet.chains[0].id,
+            type: 'account'
+          })
+        } catch (error) {
+          // unsupported network for transaction events
+        }
+      }
+    }
   })
 
-  // also when accounts change update Balance and ENS
+  // also when accounts change, update Balance and ENS
   accountsChanged$
     .pipe(
       switchMap(async ([address]) => {
@@ -163,12 +200,45 @@ export function trackWallet(
   )
 
   // Update chain on wallet when chainId changed
-  chainChanged$.subscribe(chainId => {
+  chainChanged$.subscribe(async chainId => {
     const { wallets } = state.get()
     const { chains, accounts } = wallets.find(wallet => wallet.label === label)
     const [connectedWalletChain] = chains
 
     if (chainId === connectedWalletChain.id) return
+
+    if (state.get().notify.enabled) {
+      const sdk = await getBlocknativeSdk()
+
+      if (sdk) {
+        const wallet = state
+          .get()
+          .wallets.find(wallet => wallet.label === label)
+
+        // Unsubscribe with timeout of 60 seconds
+        // to allow for any currently inflight transactions
+        wallet.accounts.forEach(({ address }) => {
+          sdk.unsubscribe({
+            id: address,
+            chainId: wallet.chains[0].id,
+            timeout: 60000
+          })
+        })
+
+        // resubscribe for new chainId
+        wallet.accounts.forEach(({ address }) => {
+          try {
+            sdk.subscribe({
+              id: address,
+              chainId: chainId,
+              type: 'account'
+            })
+          } catch (error) {
+            // unsupported network for transaction events
+          }
+        })
+      }
+    }
 
     const resetAccounts = accounts.map(
       ({ address }) =>
@@ -231,13 +301,7 @@ export async function getEns(
   // chain we don't recognize and don't have a rpcUrl for requests
   if (!chain) return null
 
-  if (!ethersProviders[chain.rpcUrl]) {
-    ethersProviders[chain.rpcUrl] = new providers.StaticJsonRpcProvider(
-      chain.rpcUrl
-    )
-  }
-
-  const provider = ethersProviders[chain.rpcUrl]
+  const provider = getProvider(chain)
 
   try {
     const name = await provider.lookupAddress(address)
@@ -277,20 +341,16 @@ export async function getBalance(
   // chain we don't recognize and don't have a rpcUrl for requests
   if (!chain) return null
 
-  if (!ethersProviders[chain.rpcUrl]) {
-    ethersProviders[chain.rpcUrl] = new providers.StaticJsonRpcProvider(
-      chain.rpcUrl
-    )
-  }
-
-  const provider = ethersProviders[chain.rpcUrl]
+  const { wallets } = state.get()
 
   try {
-    const balanceWei = await provider.getBalance(address)
-
-    return balanceWei
-      ? { [chain.token || 'eth']: utils.formatEther(balanceWei) }
-      : null
+    const wallet = wallets.find(wallet => !!wallet.provider)
+    const provider = wallet.provider
+    const balanceHex = await provider.request({
+      method: 'eth_getBalance',
+      params: [address, 'latest']
+    })
+    return balanceHex ? { [chain.token || 'eth']: weiToEth(balanceHex) } : null
   } catch (error) {
     console.error(error)
     return null
@@ -322,7 +382,10 @@ export function addNewChain(
           symbol: chain.token,
           decimals: 18
         },
-        rpcUrls: [chain.rpcUrl]
+        rpcUrls: [chain.publicRpcUrl || chain.rpcUrl],
+        blockExplorerUrls: chain.blockExplorerUrl
+          ? [chain.blockExplorerUrl]
+          : undefined
       }
     ]
   })
