@@ -1,7 +1,7 @@
-import type { Chain, WalletInit } from '@web3-onboard/common'
-import { internalState$ } from '../streams'
-import { initializeWalletModules } from '../utils'
-import { dispatch } from './index'
+import type { Chain, WalletInit, WalletModule } from '@web3-onboard/common'
+import { nanoid } from 'nanoid'
+import { dispatch } from './index.js'
+import { configuration } from '../configuration.js'
 
 import type {
   Account,
@@ -11,18 +11,37 @@ import type {
   RemoveWalletAction,
   ResetStoreAction,
   SetWalletModulesAction,
+  SetLocaleAction,
   UpdateAccountAction,
   UpdateAccountCenterAction,
   UpdateWalletAction,
-  WalletState
-} from '../types'
+  WalletState,
+  UpdateNotifyAction,
+  Notification,
+  AddNotificationAction,
+  RemoveNotificationAction,
+  UpdateAllWalletsAction,
+  CustomNotification,
+  UpdateNotification,
+  CustomNotificationUpdate,
+  Notify,
+  ConnectModalOptions,
+  UpdateConnectModalAction
+} from '../types.js'
 
 import {
   validateAccountCenterUpdate,
+  validateLocale,
+  validateNotification,
+  validateCustomNotification,
+  validateCustomNotificationUpdate,
   validateString,
   validateWallet,
-  validateWalletInit
-} from '../validation'
+  validateWalletInit,
+  validateUpdateBalances,
+  validateNotify,
+  validateConnectModalUpdate
+} from '../validation.js'
 
 import {
   ADD_CHAINS,
@@ -32,16 +51,23 @@ import {
   REMOVE_WALLET,
   UPDATE_ACCOUNT,
   UPDATE_ACCOUNT_CENTER,
-  SET_WALLET_MODULES
-} from './constants'
+  UPDATE_NOTIFY,
+  SET_WALLET_MODULES,
+  SET_LOCALE,
+  ADD_NOTIFICATION,
+  REMOVE_NOTIFICATION,
+  UPDATE_ALL_WALLETS,
+  UPDATE_CONNECT_MODAL
+} from './constants.js'
 
 export function addChains(chains: Chain[]): void {
   // chains are validated on init
   const action = {
     type: ADD_CHAINS,
-    payload: chains.map(({ namespace = 'evm', ...rest }) => ({
+    payload: chains.map(({ namespace = 'evm', id, ...rest }) => ({
       ...rest,
-      namespace
+      namespace,
+      id: id.toLowerCase()
     }))
   }
 
@@ -84,7 +110,7 @@ export function updateWallet(id: string, update: Partial<WalletState>): void {
 }
 
 export function removeWallet(id: string): void {
-  const error = validateString(id)
+  const error = validateString(id, 'wallet id')
 
   if (error) {
     throw error
@@ -98,6 +124,30 @@ export function removeWallet(id: string): void {
   }
 
   dispatch(action as RemoveWalletAction)
+}
+
+export function setPrimaryWallet(wallet: WalletState, address?: string): void {
+  const error =
+    validateWallet(wallet) || (address && validateString(address, 'address'))
+
+  if (error) {
+    throw error
+  }
+
+  // if also setting the primary account
+  if (address) {
+    const account = wallet.accounts.find(ac => ac.address === address)
+
+    if (account) {
+      wallet.accounts = [
+        account,
+        ...wallet.accounts.filter(({ address }) => address !== account.address)
+      ]
+    }
+  }
+
+  // add wallet will set it to first wallet since it already exists
+  addWallet(wallet)
 }
 
 export function updateAccount(
@@ -134,6 +184,138 @@ export function updateAccountCenter(
   dispatch(action as UpdateAccountCenterAction)
 }
 
+export function updateConnectModal(
+  update: ConnectModalOptions | Partial<ConnectModalOptions>
+): void {
+  const error = validateConnectModalUpdate(update)
+
+  if (error) {
+    throw error
+  }
+
+  const action = {
+    type: UPDATE_CONNECT_MODAL,
+    payload: update
+  }
+
+  dispatch(action as UpdateConnectModalAction)
+}
+
+export function updateNotify(update: Partial<Notify>): void {
+  const error = validateNotify(update)
+
+  if (error) {
+    throw error
+  }
+
+  const action = {
+    type: UPDATE_NOTIFY,
+    payload: update
+  }
+
+  dispatch(action as UpdateNotifyAction)
+}
+
+export function addNotification(notification: Notification): void {
+  const error = validateNotification(notification)
+
+  if (error) {
+    throw error
+  }
+
+  const action = {
+    type: ADD_NOTIFICATION,
+    payload: notification
+  }
+
+  dispatch(action as AddNotificationAction)
+}
+
+export function addCustomNotification(
+  notification: CustomNotificationUpdate
+): void {
+  const customNotificationError = validateCustomNotificationUpdate(notification)
+
+  if (customNotificationError) {
+    throw customNotificationError
+  }
+
+  const action = {
+    type: ADD_NOTIFICATION,
+    payload: notification
+  }
+
+  dispatch(action as AddNotificationAction)
+}
+
+export function customNotification(updatedNotification: CustomNotification): {
+  dismiss: () => void
+  update: UpdateNotification
+} {
+  const customNotificationError =
+    validateCustomNotification(updatedNotification)
+
+  if (customNotificationError) {
+    throw customNotificationError
+  }
+
+  const customIdKey = `customNotification-${nanoid()}`
+  const notification: CustomNotificationUpdate = {
+    ...updatedNotification,
+    id: customIdKey,
+    key: customIdKey
+  }
+  addCustomNotification(notification)
+
+  const dismiss = () => removeNotification(notification.id)
+
+  const update = (
+    notificationUpdate: CustomNotification
+  ): {
+    dismiss: () => void
+    update: UpdateNotification
+  } => {
+    const customNotificationError =
+      validateCustomNotification(updatedNotification)
+
+    if (customNotificationError) {
+      throw customNotificationError
+    }
+
+    const notificationAfterUpdate: CustomNotificationUpdate = {
+      ...notificationUpdate,
+      id: notification.id,
+      key: notification.key
+    }
+    addCustomNotification(notificationAfterUpdate)
+
+    return {
+      dismiss,
+      update
+    }
+  }
+
+  addCustomNotification(notification)
+
+  return {
+    dismiss,
+    update
+  }
+}
+
+export function removeNotification(id: Notification['id']): void {
+  if (typeof id !== 'string') {
+    throw new Error('Notification id must be of type string')
+  }
+
+  const action = {
+    type: REMOVE_NOTIFICATION,
+    payload: id
+  }
+
+  dispatch(action as RemoveNotificationAction)
+}
+
 export function resetStore(): void {
   const action = {
     type: RESET_STORE
@@ -149,15 +331,69 @@ export function setWalletModules(wallets: WalletInit[]): void {
     throw error
   }
 
-  const modules = initializeWalletModules(
-    wallets,
-    internalState$.getValue().device
-  )
+  const modules = initializeWalletModules(wallets)
+  const dedupedWallets = uniqueWalletsByLabel(modules)
 
   const action = {
     type: SET_WALLET_MODULES,
-    payload: modules
+    payload: dedupedWallets
   }
 
   dispatch(action as SetWalletModulesAction)
+}
+
+export function setLocale(locale: string): void {
+  const error = validateLocale(locale)
+
+  if (error) {
+    throw error
+  }
+
+  const action = {
+    type: SET_LOCALE,
+    payload: locale
+  }
+
+  dispatch(action as SetLocaleAction)
+}
+
+export function updateAllWallets(wallets: WalletState[]): void {
+  const error = validateUpdateBalances(wallets)
+
+  if (error) {
+    throw error
+  }
+
+  const action = {
+    type: UPDATE_ALL_WALLETS,
+    payload: wallets
+  }
+
+  dispatch(action as UpdateAllWalletsAction)
+}
+
+// ==== HELPERS ==== //
+export function initializeWalletModules(modules: WalletInit[]): WalletModule[] {
+  const { device } = configuration
+  return modules.reduce((acc, walletInit) => {
+    const initialized = walletInit({ device })
+
+    if (initialized) {
+      // injected wallets is an array of wallets
+      acc.push(...(Array.isArray(initialized) ? initialized : [initialized]))
+    }
+
+    return acc
+  }, [] as WalletModule[])
+}
+
+export function uniqueWalletsByLabel(
+  walletModuleList: WalletModule[]
+): WalletModule[] {
+  return walletModuleList.filter(
+    (wallet, i) =>
+      walletModuleList.findIndex(
+        (innerWallet: WalletModule) => innerWallet.label === wallet.label
+      ) === i
+  )
 }
