@@ -9,19 +9,14 @@ import {
   shareReplay
 } from 'rxjs/operators'
 
-import { resetStore } from './store/actions'
-import { state } from './store'
+import { resetStore } from './store/actions.js'
+import { state } from './store/index.js'
 
-import type { WalletState, InternalState, ConnectOptions } from './types'
+import type { WalletState, ConnectOptions } from './types.js'
+import type { EthereumTransactionData } from 'bnc-sdk'
 
 export const reset$ = new Subject<void>()
 export const disconnectWallet$ = new Subject<WalletState['label']>()
-
-export const internalState$ = new BehaviorSubject<InternalState>({
-  svelteInstance: null,
-  appMetadata: null,
-  device: null
-})
 
 export const connectWallet$ = new BehaviorSubject<{
   autoSelect?: ConnectOptions['autoSelect']
@@ -46,6 +41,30 @@ reset$.pipe(withLatestFrom(wallets$), pluck('1')).subscribe(wallets => {
 
   resetStore()
 })
+
+// keep transactions for all notifications for replacement actions
+export const transactions$ = new BehaviorSubject<EthereumTransactionData[]>([])
+
+export function updateTransaction(tx: EthereumTransactionData): void {
+  const currentTransactions = transactions$.getValue()
+
+  const txIndex = currentTransactions.findIndex(({ hash }) => hash === tx.hash)
+
+  if (txIndex !== -1) {
+    const updatedTransactions = currentTransactions.map((val, i) =>
+      i === txIndex ? tx : val
+    )
+
+    transactions$.next(updatedTransactions)
+  } else {
+    transactions$.next([...currentTransactions, tx])
+  }
+}
+
+export function removeTransaction(hash: string): void {
+  const currentTransactions = transactions$.getValue()
+  transactions$.next(currentTransactions.filter(tx => tx.hash !== hash))
+}
 
 export const onMount$ = defer(() => {
   const subject = new Subject<void>()
