@@ -8,33 +8,52 @@ import type {
   ChainId,
   Chain,
   WalletInit,
-  WalletModule
+  WalletModule,
+  ChainWithDecimalId
 } from '@web3-onboard/common'
 
-import ethereumIcon from './icons/ethereum'
-import polygonIcon from './icons/polygon'
-import questionIcon from './icons/question'
-import binanceIcon from './icons/binance'
-import fantomIcon from './icons/fantom'
-import optimismIcon from './icons/optimism'
-import avalancheIcon from './icons/avalanche'
-import celoIcon from './icons/celo'
-import gnosisIcon from './icons/gnosis'
-import harmonyOneIcon from './icons/harmony-one'
-import arbitrumIcon from './icons/arbitrum'
+import ethereumIcon from './icons/ethereum.js'
+import polygonIcon from './icons/polygon.js'
+import questionIcon from './icons/question.js'
+import binanceIcon from './icons/binance.js'
+import fantomIcon from './icons/fantom.js'
+import optimismIcon from './icons/optimism.js'
+import avalancheIcon from './icons/avalanche.js'
+import celoIcon from './icons/celo.js'
+import gnosisIcon from './icons/gnosis.js'
+import harmonyOneIcon from './icons/harmony-one.js'
+import arbitrumIcon from './icons/arbitrum.js'
 
-import type { ChainStyle, ConnectedChain } from './types'
+import hourglass from './icons/hourglass.js'
+import checkmark from './icons/checkmark.js'
+import error from './icons/error.js'
+import info from './icons/info.js'
 
-export function getDevice(): Device {
-  const parsed = bowser.getParser(window.navigator.userAgent)
-  const os = parsed.getOS()
-  const browser = parsed.getBrowser()
-  const { type } = parsed.getPlatform()
+import type {
+  ChainStyle,
+  ConnectedChain,
+  DeviceNotBrowser,
+  NotifyEventStyles
+} from './types.js'
 
-  return {
-    type: type as DeviceType,
-    os: os as DeviceOS,
-    browser: browser as DeviceBrowser
+export function getDevice(): Device | DeviceNotBrowser {
+  if (typeof window !== 'undefined') {
+    const parsed = bowser.getParser(window.navigator.userAgent)
+    const os = parsed.getOS()
+    const browser = parsed.getBrowser()
+    const { type } = parsed.getPlatform()
+
+    return {
+      type: type as DeviceType,
+      os: os as DeviceOS,
+      browser: browser as DeviceBrowser
+    }
+  } else {
+    return {
+      type: null,
+      os: null,
+      browser: null
+    }
   }
 }
 
@@ -65,6 +84,29 @@ export function shortenEns(ens: string): string {
   return ens.length > 11 ? `${ens.slice(0, 4)}...${ens.slice(-6)}` : ens
 }
 
+export async function copyWalletAddress(text: string): Promise<void> {
+  try {
+    const copy = await navigator.clipboard.writeText(text)
+    return copy
+  } catch (err) {
+    console.error('Failed to copy: ', err)
+  }
+}
+
+export const toHexString = (val: number | string): string =>
+  typeof val === 'number' ? `0x${val.toString(16)}` : val
+
+export function chainIdToHex(chains: Chain[] | ChainWithDecimalId[]): Chain[] {
+  return chains.map(({ id, ...rest }) => {
+    const hexId = toHexString(id)
+    return { id: hexId, ...rest }
+  })
+}
+
+export function gweiToWeiHex(gwei: number): string {
+  return `0x${(gwei * 1e9).toString(16)}`
+}
+
 export const chainIdToLabel: Record<string, string> = {
   '0x1': 'Ethereum',
   '0x3': 'Ropsten',
@@ -75,11 +117,25 @@ export const chainIdToLabel: Record<string, string> = {
   '0x89': 'Polygon',
   '0xfa': 'Fantom',
   '0xa': 'Optimism',
+  '0x45': 'Optimism Kovan',
   '0xa86a': 'Avalanche',
   '0xa4ec': 'Celo',
   '0x64': 'Gnosis',
   '0x63564C40': 'Harmony One',
   '0xa4b1': 'Arbitrum'
+}
+
+export const networkToChainId: Record<string, ChainId> = {
+  main: '0x1',
+  ropsten: '0x3',
+  rinkeby: '0x4',
+  goerli: '0x5',
+  kovan: '0x2a',
+  xdai: '0x64',
+  'bsc-main': '0x38',
+  'matic-main': '0x89',
+  'fantom-main': '0xfa',
+  'matic-mumbai': '0x80001'
 }
 
 export const chainStyles: Record<string, ChainStyle> = {
@@ -116,6 +172,10 @@ export const chainStyles: Record<string, ChainStyle> = {
     color: '#1969FF'
   },
   '0xa': {
+    icon: optimismIcon,
+    color: '#FF0420'
+  },
+  '0x45': {
     icon: optimismIcon,
     color: '#FF0420'
   },
@@ -173,3 +233,30 @@ export function initializeWalletModules(
     return acc
   }, [] as WalletModule[])
 }
+
+export const defaultNotifyEventStyles: Record<string, NotifyEventStyles> = {
+  pending: {
+    backgroundColor: 'var(--onboard-primary-700, var(--primary-700))',
+    borderColor: '#6370E5',
+    eventIcon: hourglass
+  },
+  success: {
+    backgroundColor: '#052E17',
+    borderColor: 'var(--onboard-success-300, var(--success-300))',
+    eventIcon: checkmark
+  },
+  error: {
+    backgroundColor: '#FDB1B11A',
+    borderColor: 'var(--onboard-danger-300, var(--danger-300))',
+    eventIcon: error
+  },
+  hint: {
+    backgroundColor: 'var(--onboard-gray-500, var(--gray-500))',
+    borderColor: 'var(--onboard-gray-500, var(--gray-500))',
+    iconColor: 'var(--onboard-gray-100, var(--gray-100))',
+    eventIcon: info
+  }
+}
+
+export const wait = (time: number): Promise<void> =>
+  new Promise(resolve => setTimeout(resolve, time))
