@@ -1,5 +1,16 @@
 import Joi from 'joi'
-import type { ChainId, WalletInit, WalletModule } from '@web3-onboard/common'
+
+import {
+  type ChainId,
+  type DecimalChainId,
+  type WalletInit,
+  type WalletModule,
+  type ValidateReturn,
+  chainNamespaceValidation,
+  chainIdValidation,
+  chainValidation,
+  validate
+} from '@web3-onboard/common'
 
 import type {
   InitOptions,
@@ -7,27 +18,22 @@ import type {
   ConnectOptions,
   DisconnectOptions,
   ConnectOptionsString,
-  AccountCenter
-} from './types'
+  AccountCenter,
+  TransactionHandlerReturn,
+  NotifyOptions,
+  Notification,
+  CustomNotification,
+  CustomNotificationUpdate,
+  Notify,
+  PreflightNotificationsOptions,
+  ConnectModalOptions
+} from './types.js'
 
-const chainId = Joi.string().pattern(/^0x[0-9a-fA-F]+$/)
-const chainNamespace = Joi.string().valid('evm')
 const unknownObject = Joi.object().unknown()
-// const address = Joi.string().regex(/^0x[a-fA-F0-9]{40}$/)
-
-const chain = Joi.object({
-  namespace: chainNamespace,
-  id: chainId.required(),
-  rpcUrl: Joi.string().required(),
-  label: Joi.string().required(),
-  token: Joi.string().required(),
-  icon: Joi.string(),
-  color: Joi.string()
-})
 
 const connectedChain = Joi.object({
-  namespace: chainNamespace.required(),
-  id: chainId.required()
+  namespace: chainNamespaceValidation.required(),
+  id: chainIdValidation.required()
 })
 
 const ens = Joi.any().allow(
@@ -47,13 +53,13 @@ const balance = Joi.any().allow(
   null
 )
 
-const account = {
+const account = Joi.object({
   address: Joi.string().required(),
   ens,
   balance
-}
+})
 
-const chains = Joi.array().items(chain)
+const chains = Joi.array().items(chainValidation)
 const accounts = Joi.array().items(account)
 
 const wallet = Joi.object({
@@ -64,6 +70,10 @@ const wallet = Joi.object({
   accounts,
   chains: Joi.array().items(connectedChain)
 })
+  .required()
+  .error(new Error('wallet must be defined'))
+
+const wallets = Joi.array().items(wallet)
 
 const recommendedWallet = Joi.object({
   name: Joi.string().required(),
@@ -97,38 +107,81 @@ const walletModule = Joi.object({
 
 const walletInit = Joi.array().items(Joi.function()).required()
 
-const accountCenterPosition = Joi.string().valid(
+const locale = Joi.string()
+
+const commonPositions = Joi.string().valid(
   'topRight',
   'bottomRight',
   'bottomLeft',
   'topLeft'
 )
 
+const gasPriceProbabilities = [70, 80, 90, 95, 99]
+
+const notify = Joi.object({
+  transactionHandler: Joi.function(),
+  enabled: Joi.boolean(),
+  position: commonPositions,
+  replacement: Joi.object({
+    gasPriceProbability: Joi.object({
+      speedup: Joi.number().valid(...gasPriceProbabilities),
+      cancel: Joi.number().valid(...gasPriceProbabilities)
+    })
+  })
+})
+
+const notifyOptions = Joi.object({
+  desktop: notify,
+  mobile: notify
+})
+
+const accountCenterInitOptions = Joi.object({
+  enabled: Joi.boolean(),
+  position: commonPositions,
+  minimal: Joi.boolean(),
+  containerElement: Joi.string()
+})
+
+const accountCenter = Joi.object({
+  enabled: Joi.boolean(),
+  position: commonPositions,
+  expanded: Joi.boolean(),
+  minimal: Joi.boolean(),
+  containerElement: Joi.string()
+})
+
+const connectModalOptions = Joi.object({
+  showSidebar: Joi.boolean()
+})
+
 const initOptions = Joi.object({
   wallets: walletInit,
   chains: chains.required(),
   appMetadata: appMetadata,
   i18n: Joi.object().unknown(),
+  apiKey: Joi.string(),
   accountCenter: Joi.object({
-    desktop: Joi.object({
-      enabled: Joi.boolean(),
-      position: accountCenterPosition
-    }),
-    mobile: Joi.object({
-      enabled: Joi.boolean(),
-      position: accountCenterPosition
-    })
-  })
+    desktop: accountCenterInitOptions,
+    mobile: accountCenterInitOptions
+  }),
+  notify: [notifyOptions, notify],
+  gas: Joi.object({
+    get: Joi.function().required(),
+    stream: Joi.function().required()
+  }),
+  connect: connectModalOptions
 })
 
 const connectOptions = Joi.object({
-  autoSelect: [
-    Joi.object({
-      label: Joi.string().required(),
-      disableModals: Joi.boolean()
-    }),
-    Joi.string()
-  ]
+  autoSelect: Joi.alternatives()
+    .try(
+      Joi.object({
+        label: Joi.string().required(),
+        disableModals: Joi.boolean()
+      }),
+      Joi.string()
+    )
+    .required()
 })
 
 const disconnectOptions = Joi.object({
@@ -136,23 +189,63 @@ const disconnectOptions = Joi.object({
 }).required()
 
 const setChainOptions = Joi.object({
-  chainId: chainId.required(),
-  chainNamespace: chainNamespace,
+  chainId: chainIdValidation.required(),
+  chainNamespace: chainNamespaceValidation,
   wallet: Joi.string()
 })
 
-const accountCenter = Joi.object({
-  enabled: Joi.boolean(),
-  position: accountCenterPosition,
-  expanded: Joi.boolean()
+const customNotificationUpdate = Joi.object({
+  key: Joi.string().required(),
+  type: Joi.string().allow('pending', 'error', 'success', 'hint'),
+  eventCode: Joi.string(),
+  message: Joi.string().required(),
+  id: Joi.string().required(),
+  autoDismiss: Joi.number(),
+  onClick: Joi.function(),
+  link: Joi.string()
 })
 
-type ValidateReturn = Joi.ValidationResult | null
+const preflightNotifications = Joi.object({
+  sendTransaction: Joi.function(),
+  estimateGas: Joi.function(),
+  gasPrice: Joi.function(),
+  balance: Joi.alternatives(Joi.string(), Joi.number()),
+  txDetails: Joi.object({
+    value: Joi.alternatives(Joi.string(), Joi.number()),
+    to: Joi.string(),
+    from: Joi.string()
+  }),
+  txApproveReminderTimeout: Joi.number()
+})
 
-function validate(validator: Joi.Schema, data: unknown): ValidateReturn {
-  const result = validator.validate(data)
-  return result.error ? result : null
-}
+const customNotification = Joi.object({
+  key: Joi.string(),
+  type: Joi.string().allow('pending', 'error', 'success', 'hint'),
+  eventCode: Joi.string(),
+  message: Joi.string(),
+  id: Joi.string(),
+  autoDismiss: Joi.number(),
+  onClick: Joi.function(),
+  link: Joi.string()
+})
+
+const notification = Joi.object({
+  id: Joi.string().required(),
+  key: Joi.string().required(),
+  type: Joi.string().allow('pending', 'error', 'success', 'hint').required(),
+  eventCode: Joi.string().required(),
+  message: Joi.string().required(),
+  autoDismiss: Joi.number().required(),
+  network: Joi.string().required(),
+  startTime: Joi.number(),
+  onClick: Joi.function(),
+  link: Joi.string()
+})
+
+const transactionHandlerReturn = Joi.any().allow(
+  customNotificationUpdate,
+  Joi.boolean().allow(false)
+)
 
 export function validateWallet(
   data: WalletState | Partial<WalletState>
@@ -180,12 +273,18 @@ export function validateDisconnectOptions(
   return validate(disconnectOptions, data)
 }
 
-export function validateString(str: string): ValidateReturn {
-  return validate(Joi.string().required(), str)
+export function validateString(str: string, label?: string): ValidateReturn {
+  return validate(
+    Joi.string()
+      .required()
+      .label(label || 'value'),
+    str
+  )
 }
 
 export function validateSetChainOptions(data: {
-  chainId: ChainId
+  chainId: ChainId | DecimalChainId
+  chainNamespace?: string
   wallet?: WalletState['label']
 }): ValidateReturn {
   return validate(setChainOptions, data)
@@ -197,6 +296,57 @@ export function validateAccountCenterUpdate(
   return validate(accountCenter, data)
 }
 
+export function validateConnectModalUpdate(
+  data: ConnectModalOptions | Partial<ConnectModalOptions>
+): ValidateReturn {
+  return validate(connectModalOptions, data)
+}
+
 export function validateWalletInit(data: WalletInit[]): ValidateReturn {
   return validate(walletInit, data)
+}
+
+export function validateLocale(data: string): ValidateReturn {
+  return validate(locale, data)
+}
+
+export function validateNotify(data: Partial<Notify>): ValidateReturn {
+  return validate(notify, data)
+}
+
+export function validateNotifyOptions(
+  data: Partial<NotifyOptions>
+): ValidateReturn {
+  return validate(notifyOptions, data)
+}
+
+export function validateTransactionHandlerReturn(
+  data: TransactionHandlerReturn
+): ValidateReturn {
+  return validate(transactionHandlerReturn, data)
+}
+
+export function validateNotification(data: Notification): ValidateReturn {
+  return validate(notification, data)
+}
+export function validatePreflightNotifications(
+  data: PreflightNotificationsOptions
+): ValidateReturn {
+  return validate(preflightNotifications, data)
+}
+
+export function validateCustomNotificationUpdate(
+  data: CustomNotificationUpdate
+): ValidateReturn {
+  return validate(customNotificationUpdate, data)
+}
+
+export function validateCustomNotification(
+  data: CustomNotification
+): ValidateReturn {
+  return validate(customNotification, data)
+}
+
+export function validateUpdateBalances(data: WalletState[]): ValidateReturn {
+  return validate(wallets, data)
 }
