@@ -1,20 +1,106 @@
 <script lang="ts">
   import { shareReplay, startWith } from 'rxjs/operators'
-  import { connectWallet$, switchChainModal$, wallets$ } from '../streams'
-  import { state } from '../store'
+  import { connectWallet$, switchChainModal$, wallets$ } from '../streams.js'
+  import { state } from '../store/index.js'
   import Connect from './connect/Index.svelte'
   import SwitchChain from './chain/SwitchChain.svelte'
   import ActionRequired from './connect/ActionRequired.svelte'
-  import AccountCenter from './account-center/Index.svelte'
+  import { configuration } from '../configuration.js'
+  import type { Observable } from 'rxjs'
+  import type { Notification } from '../types.js'
 
+  const { device } = configuration
   const accountCenter$ = state
     .select('accountCenter')
     .pipe(startWith(state.get().accountCenter), shareReplay(1))
+
+  const notify$ = state
+    .select('notify')
+    .pipe(startWith(state.get().notify), shareReplay(1))
+
+  const notifications$: Observable<Notification[]> = state
+    .select('notifications')
+    .pipe(startWith(state.get().notifications))
+
+  const accountCenterPositioning = 'account-center'
+  const notifyPositioning = 'notify-onboard-container'
+  const setPositioningDefaults = (targetComponentVariable: string) => {
+    return {
+      topLeft: `
+        top: var(--${targetComponentVariable}-position-top, 0); 
+        left: var(--${targetComponentVariable}-position-left, 0);`,
+      topRight: `
+        top: var(--${targetComponentVariable}-position-top, 0); 
+        right: var(--${targetComponentVariable}-position-right, 0);`,
+      bottomRight: `
+        bottom: var(--${targetComponentVariable}-position-bottom, 0); 
+        right: var(--${targetComponentVariable}-position-right, 0);`,
+      bottomLeft: `
+        bottom: var(--${targetComponentVariable}-position-bottom, 0); 
+        left: var(--${targetComponentVariable}-position-left, 0);`
+    }
+  }
+
+  const accountCenterComponent = $accountCenter$.enabled
+    ? import('./account-center/Index.svelte').then(mod => mod.default)
+    : Promise.resolve(null)
+
+  const notifyComponent = $notify$.enabled
+    ? import('./notify/Index.svelte').then(mod => mod.default)
+    : Promise.resolve(null)
+
+  $: sharedContainer =
+    $accountCenter$.enabled &&
+    $notify$.enabled &&
+    $notify$.position === $accountCenter$.position
+
+  $: samePositionOrMobile =
+    device.type === 'mobile' || $accountCenter$.position === $notify$.position
+
+  $: sharedMobileContainerCheck =
+    device.type === 'mobile' &&
+    (($notify$.position.includes('bottom') &&
+      $accountCenter$.position.includes('bottom')) ||
+      ($notify$.position.includes('top') &&
+        $accountCenter$.position.includes('top')))
+
+  $: separateMobileContainerCheck =
+    device.type === 'mobile' &&
+    (($notify$.position.includes('top') &&
+      $accountCenter$.position.includes('bottom')) ||
+      ($notify$.position.includes('bottom') &&
+        $accountCenter$.position.includes('top')))
+
+  $: displayNotifySeparate =
+    $notify$.enabled &&
+    (!$accountCenter$.enabled ||
+      ($notify$.position !== $accountCenter$.position &&
+        device.type !== 'mobile') ||
+      separateMobileContainerCheck ||
+      !$wallets$.length)
+
+  $: displayAccountCenterSeparate =
+    $accountCenter$.enabled &&
+    (!$notify$.enabled ||
+      ($notify$.position !== $accountCenter$.position &&
+        device.type !== 'mobile') ||
+      separateMobileContainerCheck) &&
+    $wallets$.length
+
+  $: displayAccountCenterNotifySameContainer =
+    $notify$.enabled &&
+    $accountCenter$.enabled &&
+    $wallets$.length &&
+    (sharedContainer || sharedMobileContainerCheck)
 </script>
 
 <style>
   :global(.flex) {
     display: flex;
+  }
+
+  :global(.inline-flex) {
+    display: inline-flex;
   }
 
   :global(.flex-column) {
@@ -43,6 +129,10 @@
 
   :global(.justify-between) {
     justify-content: space-between;
+  }
+
+  :global(.justify-end) {
+    justify-content: flex-end;
   }
 
   :global(.justify-around) {
@@ -96,7 +186,7 @@
   :global(input[type='checkbox']) {
     -webkit-appearance: none;
     width: auto;
-    background-color: var(--onboard-white, var(--white));
+    background: var(--onboard-white, var(--white));
     outline: 1px solid var(--onboard-gray-300, var(--gray-300));
     border: none;
     padding: 0.5em;
@@ -116,7 +206,7 @@
   }
 
   :global(input[type='checkbox']:checked) {
-    background-color: var(
+    background: var(
       --onboard-checkbox-background,
       var(--onboard-primary-500, var(--primary-500))
     );
@@ -152,7 +242,7 @@
   }
 
   :global(input:disabled, textarea:disabled, select:disabled) {
-    background-color: var(--gray-100);
+    background: var(--gray-100);
   }
 
   :global(input::-moz-focus-inner) {
@@ -200,14 +290,14 @@
   :global(.button-neutral-solid) {
     width: 100%;
     border-radius: 8px;
-    background-color: var(--onboard-gray-500, var(--gray-500));
+    background: var(--onboard-gray-500, var(--gray-500));
     color: var(--onboard-white, var(--white));
     line-height: var(--onboard-font-line-height-3, var(--font-line-height-3));
   }
 
   :global(.button-neutral-solid-b) {
     width: 100%;
-    background-color: var(--onboard-gray-100, var(--gray-100));
+    background: var(--onboard-gray-100, var(--gray-100));
     color: var(--onboard-gray-500, var(--gray-500));
     line-height: var(--onboard-font-line-height-3, var(--font-line-height-3));
   }
@@ -217,10 +307,10 @@
   }
 
   :global(.button-neutral-solid:hover) {
-    background-color: var(--onboard-gray-700, var(--gray-700));
+    background: var(--onboard-gray-700, var(--gray-700));
   }
   :global(.button-neutral-solid-b:hover) {
-    background-color: var(--onboard-gray-200, var(--gray-200));
+    background: var(--onboard-gray-200, var(--gray-200));
   }
 
   :global(.button-neutral-solid:active) {
@@ -229,7 +319,25 @@
 
   :global(.button-neutral-solid-b:active) {
     color: var(--onboard-gray-600, var(--gray-600));
-    background-color: var(--onboard-gray-300, var(--gray-300));
+    background: var(--onboard-gray-300, var(--gray-300));
+  }
+
+  .container {
+    padding: 16px;
+    font-family: var(--onboard-font-family-normal, var(--font-family-normal));
+    width: 100%;
+    pointer-events: none;
+    touch-action: none;
+  }
+
+  .z-indexed {
+    z-index: var(--account-center-z-index);
+  }
+
+  @media all and (min-width: 428px) {
+    .container {
+      max-width: 348px;
+    }
   }
 </style>
 
@@ -245,6 +353,112 @@
   <SwitchChain />
 {/if}
 
-{#if $accountCenter$.enabled && $wallets$.length}
-  <AccountCenter settings={$accountCenter$} />
+{#if displayAccountCenterNotifySameContainer}
+  <div
+    class="container flex flex-column fixed z-indexed"
+    style="{setPositioningDefaults(accountCenterPositioning)[
+      $accountCenter$.position
+    ]}; {device.type === 'mobile' && $accountCenter$.position.includes('top')
+      ? 'padding-bottom: 0;'
+      : device.type === 'mobile' && $accountCenter$.position.includes('bottom')
+      ? 'padding-top:0;'
+      : ''} "
+  >
+    {#if $notify$.position.includes('bottom') && $accountCenter$.position.includes('bottom') && samePositionOrMobile}
+      {#await notifyComponent then Notify}
+        {#if Notify}
+          <svelte:component
+            this={Notify}
+            notifications={$notifications$}
+            position={$notify$.position}
+            {sharedContainer}
+          />
+        {/if}
+      {/await}
+    {/if}
+    <div
+      style={!$accountCenter$.expanded &&
+      $accountCenter$.minimal &&
+      $accountCenter$.position.includes('Right')
+        ? 'margin-left: auto'
+        : !$accountCenter$.expanded &&
+          $accountCenter$.minimal &&
+          $accountCenter$.position.includes('Left')
+        ? 'margin-right: auto'
+        : ''}
+    >
+      {#await accountCenterComponent then AccountCenter}
+        {#if AccountCenter}
+          <svelte:component this={AccountCenter} settings={$accountCenter$} />
+        {/if}
+      {/await}
+    </div>
+    {#if $notify$.position.includes('top') && $accountCenter$.position.includes('top') && samePositionOrMobile}
+      {#await notifyComponent then Notify}
+        {#if Notify}
+          <svelte:component
+            this={Notify}
+            notifications={$notifications$}
+            position={$notify$.position}
+            {sharedContainer}
+          />
+        {/if}
+      {/await}
+    {/if}
+  </div>
+{/if}
+{#if displayAccountCenterSeparate}
+  <div
+    class="container flex flex-column fixed z-indexed"
+    style="{setPositioningDefaults(accountCenterPositioning)[
+      $accountCenter$.position
+    ]}; {device.type === 'mobile' && $accountCenter$.position.includes('top')
+      ? 'padding-bottom: 0;'
+      : device.type === 'mobile' && $accountCenter$.position.includes('bottom')
+      ? 'padding-top:0;'
+      : ''} "
+  >
+    <div
+      style={!$accountCenter$.expanded &&
+      $accountCenter$.minimal &&
+      $accountCenter$.position.includes('Right')
+        ? 'margin-left: auto'
+        : !$accountCenter$.expanded &&
+          $accountCenter$.minimal &&
+          $accountCenter$.position.includes('Left')
+        ? 'margin-right: auto'
+        : ''}
+    >
+      {#if $accountCenter$.enabled && $wallets$.length}
+        {#await accountCenterComponent then AccountCenter}
+          {#if AccountCenter}
+            <svelte:component this={AccountCenter} settings={$accountCenter$} />
+          {/if}
+        {/await}
+      {/if}
+    </div>
+  </div>
+{/if}
+{#if displayNotifySeparate}
+  <div
+    class="container flex flex-column fixed z-indexed"
+    style="{setPositioningDefaults(notifyPositioning)[
+      $notify$.position
+    ]}; {device.type === 'mobile' && $notify$.position.includes('top')
+      ? 'padding-bottom: 0;'
+      : device.type === 'mobile' && $notify$.position.includes('bottom')
+      ? 'padding-top:0;'
+      : ''} "
+  >
+    {#await notifyComponent then Notify}
+      {#if Notify}
+        <svelte:component
+          this={Notify}
+          notifications={$notifications$}
+          position={$notify$.position}
+          {sharedContainer}
+        />
+      {/if}
+    {/await}
+  </div>
 {/if}
