@@ -15,7 +15,39 @@
 
 <script lang="ts">
   import { fade } from 'svelte/transition'
+  import { onDestroy, onMount } from 'svelte'
+  import { configuration } from '../../configuration.js'
 
+  const { device } = configuration
+
+  const body = document.body
+  const html = document.documentElement
+  const trackYScrollPosition = () => {
+    document.documentElement.style.setProperty(
+      '--scroll-y',
+      `${window.scrollY}px`
+    )
+  }
+
+  onMount(() => {
+    window.addEventListener('scroll', trackYScrollPosition, { passive: true })
+    const scrollY = html.style.getPropertyValue('--scroll-y')
+    device.type === 'mobile'
+      ? (html.style.position = 'fixed')
+      : (html.style.overflow = 'hidden')
+
+    body.style.top = `-${scrollY}`
+  })
+
+  onDestroy(() => {
+    device.type === 'mobile'
+      ? (html.style.position = '')
+      : (html.style.overflow = 'auto')
+    const scrollY = body.style.top
+    body.style.top = ''
+    window.scrollTo(0, parseInt(scrollY || '0') * -1)
+    window.removeEventListener('scroll', trackYScrollPosition)
+  })
   export let close: () => void
 </script>
 
@@ -48,7 +80,7 @@
   .modal-overflow {
     overflow: hidden;
   }
-  
+
   .modal-styling {
     border-radius: var(--onboard-modal-border-radius, var(--border-radius-1));
     box-shadow: var(--onboard-modal-box-shadow, var(--box-shadow-0));
@@ -57,7 +89,8 @@
   .modal {
     border-radius: var(--onboard-modal-border-radius, var(--border-radius-1));
     overflow-y: auto;
-    background: white;
+    background: var(--onboard-modal-background, white);
+    color: var(--onboard-modal-color, initial);
   }
 
   @media all and (max-width: 520px) {
@@ -71,6 +104,7 @@
 
     .modal {
       width: 100%;
+      margin: 0 16px;
     }
   }
 </style>
