@@ -1,6 +1,6 @@
 <script lang="ts">
   import { fade } from 'svelte/transition'
-  import { isSVG } from '../../utils'
+  import { isSVG } from '../../utils.js'
   import Spinner from './Spinner.svelte'
   export let size: number // px
   export let icon: Promise<string> | string // svg string or url string
@@ -78,27 +78,27 @@
   }
 
   .background-gray {
-    background: var(--onboard-gray-500, var(--gray-500));
+    background: var(--onboard-wallet-app-icon-background-gray, var(--onboard-gray-500, var(--gray-500)));
   }
 
   .background-light-gray {
-    background: var(--onboard-gray-100, var(--gray-100));
+    background: var(--onboard-wallet-app-icon-background-light-gray, var(--onboard-gray-100, var(--gray-100)));
   }
 
   .background-light-blue {
-    background: var(--onboard-primary-100, var(--primary-100));
+    background: var(--onboard-wallet-app-icon-background-light-blue, var(--onboard-primary-100, var(--primary-100)));
   }
 
   .background-green {
-    background: var(--onboard-success-100, var(--success-100));
+    background: var(--onboard-wallet-app-icon-background-green, var(--onboard-success-100, var(--success-100)));
   }
 
   .background-white {
-    background: var(--onboard-white, var(--white));
+    background: var(--onboard-wallet-app-icon-background-white, var(--onboard-white, var(--white)));
   }
 
   .background-transparent {
-    background: transparent;
+    background: var(--onboard-wallet-app-icon-background-transparent, transparent);
   }
 
   @keyframes pulse {
