@@ -12,17 +12,22 @@
   import walletConnectModule from '@web3-onboard/walletconnect'
   import coinbaseModule from '@web3-onboard/coinbase'
   import magicModule from '@web3-onboard/magic'
-  import { verifyMessage, verifyTypedData } from 'ethers/lib/utils'
+  import web3authModule from '@web3-onboard/web3auth'
+  import gas from '@web3-onboard/gas'
+  import dcentModule from '@web3-onboard/dcent'
+  import sequenceModule from '@web3-onboard/sequence'
+  import tallyHoModule from '@web3-onboard/tallyho'
+  import {
+    recoverAddress,
+    arrayify,
+    hashMessage,
+    verifyTypedData
+  } from 'ethers/lib/utils'
+  import { ethers } from 'ethers'
   import { share } from 'rxjs/operators'
   import VConsole from 'vconsole'
-  import blocknativeIcon from './blocknative-icon'
-  import blocknativeLogo from './blocknative-logo'
-
-  const toHex = text =>
-    text
-      .split('')
-      .map(c => c.charCodeAt(0).toString(16).padStart(2, '0'))
-      .join('')
+  import blocknativeIcon from './blocknative-icon.js'
+  import blocknativeLogo from './blocknative-logo.js'
 
   if (window.innerWidth < 700) {
     new VConsole()
@@ -59,7 +64,9 @@
 
   const coinbaseWallet = coinbaseModule()
 
-  const walletConnect = walletConnectModule()
+  const walletConnect = walletConnectModule({
+    connectFirstChainId: true
+  })
   const portis = portisModule({
     apiKey: 'b2b7586f-2b1e-4c30-a7fb-c2d1533b153b'
   })
@@ -68,11 +75,17 @@
     apiKey: 'pk_test_886ADCAB855632AA'
   })
 
+  const web3auth = web3authModule({
+    clientId:
+      'DJuUOKvmNnlzy6ruVgeWYWIMKLRyYtjYa9Y10VCeJzWZcygDlrYLyXsBQjpJ2hxlBO9dnl8t9GmAC2qOP5vnIGo'
+  })
+
   const torus = torusModule()
   const ledger = ledgerModule()
   const keepkey = keepkeyModule()
   const keystone = keystoneModule()
   const gnosis = gnosisModule()
+  const tallyho = tallyHoModule()
 
   const trezorOptions = {
     email: 'test@test.com',
@@ -81,54 +94,69 @@
   const trezor = trezorModule(trezorOptions)
 
   const magic = magicModule({
-    apiKey: 'pk_live_02207D744E81C2BA',
-    // userEmail: 'test@test.com' 
+    apiKey: 'pk_live_02207D744E81C2BA'
+    // userEmail: 'test@test.com'
     // userEmail is optional - if user has already logged in and/or session is still active a login modal will not appear
     // for more info see the @web3-onboard/magic docs
   })
 
+  const dcent = dcentModule()
+
+  const sequence = sequenceModule()
+
   const onboard = Onboard({
     wallets: [
+      injected,
+      web3auth,
       ledger,
       trezor,
       walletConnect,
       keepkey,
       keystone,
       coinbaseWallet,
-      injected,
       magic,
       fortmatic,
       portis,
       torus,
-      gnosis
+      gnosis,
+      dcent,
+      sequence,
+      tallyho
     ],
+    gas,
     chains: [
       {
         id: '0x1',
         token: 'ETH',
         label: 'Ethereum',
-        rpcUrl: 'https://mainnet.infura.io/v3/ababf9851fd845d0a167825f97eeb12b'
+        rpcUrl: 'https://mainnet.infura.io/v3/17c1e1500e384acfb6a72c5d2e67742e'
       },
       {
-        id: '0x3',
+        id: 3,
         token: 'tROP',
         label: 'Ropsten',
-        rpcUrl: 'https://ropsten.infura.io/v3/ababf9851fd845d0a167825f97eeb12b'
+        rpcUrl: 'https://ropsten.infura.io/v3/17c1e1500e384acfb6a72c5d2e67742e'
       },
       {
-        id: '0x4',
+        id: 4,
         token: 'rETH',
         label: 'Rinkeby',
-        rpcUrl: 'https://rinkeby.infura.io/v3/ababf9851fd845d0a167825f97eeb12b'
+        rpcUrl: 'https://rinkeby.infura.io/v3/17c1e1500e384acfb6a72c5d2e67742e'
       },
       {
-        id: '0x89',
+        id: 137,
         token: 'MATIC',
         label: 'Polygon',
         rpcUrl: 'https://matic-mainnet.chainstacklabs.com'
       },
       {
-        id: '0xa',
+        id: '0x13881',
+        token: 'MATIC',
+        label: 'Polygon - Mumbai',
+        rpcUrl: 'https://matic-mumbai.chainstacklabs.com	'
+      },
+      {
+        id: 10,
         token: 'OETH',
         label: 'Optimism',
         rpcUrl: 'https://mainnet.optimism.io'
@@ -150,43 +178,139 @@
       },
       gettingStartedGuide: 'https://blocknative.com',
       explore: 'https://blocknative.com'
-    }
-    // example customizing account center
-    // accountCenter: {
-    //   desktop: {
-    //     position: 'bottomRight'
-    //   }
-    // }
+    },
+    // // example customizing account center
+    accountCenter: {
+      desktop: {
+        position: 'topRight',
+        enabled: true,
+        minimal: false
+      }
+    },
     // example customizing copy
-    // i18n: {
-    //   en: {
-    //     connect: {
-    //       selectingWallet: {
-    //         header: 'custom text header'
-    //       }
-    //     }
-    //   }
-    // }
+    i18n: {
+      en: {
+        notify: {
+          watched: {
+            // "txConfirmed": "you paid a foo {formattedValue} {asset}!"
+          }
+        }
+      }
+    },
+    notify: {
+      desktop: {
+        enabled: true,
+        transactionHandler: transaction => {
+          console.log({ transaction })
+          //   if (transaction.eventCode === 'txConfirmed') {
+          //     return {
+          //       type: 'error',
+          //       message: 'Your in the pool, hope you brought a towel!',
+          //       autoDismiss: 0,
+          //       id: '123',
+          //       key: '321',
+          //       onClick: () =>
+          //         window.open(`https://rinkeby.etherscan.io/tx/${transaction.hash}`)
+          //     }
+          //   }
+          // if (transaction.eventCode === 'txPool') {
+          //   return {
+          //     type: 'hint',
+          //     message: 'Your in the pool, hope you brought a towel!',
+          //     autoDismiss: 0,
+          //     link: `https://ropsten.etherscan.io/tx/${transaction.hash}`
+          //   }
+          // }
+        },
+        position: 'topRight'
+      }
+    },
+    // Sign up for your free api key at www.Blocknative.com
+    apiKey: 'xxxxxx-bf21-42ec-a093-9d37e426xxxx'
   })
 
   // Subscribe to wallet updates
   const wallets$ = onboard.state.select('wallets').pipe(share())
 
-  const signTransactionMessage = provider => {
-    provider.request({
-      method: 'eth_signTransaction',
-      params: [JSON.parse(transactionObject)]
+  const signTransactionMessage = async provider => {
+    const ethersProvider = new ethers.providers.Web3Provider(provider, 'any')
+
+    const signer = ethersProvider.getSigner()
+
+    const signature = await signer.signTransaction({
+      to: '',
+      value: 100000000000000
     })
+
+    console.log(signature)
+  }
+
+  let toAddress
+  const sendTransaction = async provider => {
+    const ethersProvider = new ethers.providers.Web3Provider(provider, 'any')
+
+    const signer = ethersProvider.getSigner()
+
+    const txn = await signer.sendTransaction({
+      to: toAddress,
+      value: 100000000000000
+    })
+
+    const receipt = await txn.wait()
+    console.log(receipt)
+  }
+
+  const sendTransactionWithPreFlight = async (provider, balance) => {
+    const balanceValue = Object.values(balance)[0]
+    const ethersProvider = new ethers.providers.Web3Provider(provider, 'any')
+
+    const signer = ethersProvider.getSigner()
+    const txDetails = {
+      to: toAddress,
+      value: 100000000000000
+    }
+
+    const sendTransaction = () => {
+      return signer.sendTransaction(txDetails).then(tx => tx.hash)
+    }
+
+    const gasPrice = () =>
+      ethersProvider.getGasPrice().then(res => res.toString())
+
+    const estimateGas = () => {
+      return ethersProvider.estimateGas(txDetails).then(res => res.toString())
+    }
+
+    const transactionHash = await onboard.state.actions.preflightNotifications({
+      sendTransaction,
+      gasPrice,
+      estimateGas,
+      balance: balanceValue,
+      txDetails: txDetails
+    })
+
+    console.log(transactionHash)
   }
 
   const signMessage = async (provider, address) => {
-    const signature = await provider.request({
-      method: 'eth_sign',
-      params: [address, toHex(signMsg)]
-    })
+    const ethersProvider = new ethers.providers.Web3Provider(provider, 'any')
 
-    const recoveredAddress = verifyMessage(signMsg, signature)
-    console.log({ signMsg, signature, recoveredAddress })
+    const signer = ethersProvider?.getSigner()
+    const addr = await signer?.getAddress()
+    const signature = await signer?.signMessage(signMsg)
+
+    const recoveredAddress = recoverAddress(
+      arrayify(hashMessage(signMsg)),
+      signature
+    )
+
+    if (recoveredAddress !== address) {
+      console.error(
+        "Signature failed. Recovered address doesn' match signing address."
+      )
+    }
+
+    console.log({ signMsg, signature, recoveredAddress, addr })
   }
 
   const signTypedMessage = async (provider, address) => {
@@ -203,6 +327,10 @@
 </script>
 
 <style>
+  button {
+    width: 14rem;
+    margin: 8px;
+  }
   .connected-wallet {
     padding: 1rem;
     border-radius: 4px;
@@ -215,13 +343,14 @@
     align-items: center;
   }
 
-  /* CUSTOMIZE CSS VARIABLES */
-  :root {
-    /* --onboard-gray-100: pink; */
+  .account-info div {
+    overflow: hidden;
+    white-space: nowrap;
+    text-overflow: ellipsis;
   }
 
   .text-input {
-    width: 24rem;
+    width: 18rem;
   }
 
   .sign-transaction {
@@ -234,22 +363,98 @@
     height: 12rem;
     margin: 0;
   }
+  .notify-chain-container {
+    display: flex;
+    flex-wrap: wrap;
+  }
+  .switch-chain-container,
+  .notify-action-container {
+    display: flex;
+    flex-direction: column;
+    width: 15rem;
+  }
 </style>
 
 <main>
-  <button on:click={() => onboard.connectWallet()}>Connect Wallet</button>
+  <div class="cta">
+    <button on:click={() => onboard.connectWallet()}>Connect Wallet</button>
 
-  {#if $wallets$}
-    <button on:click={() => onboard.setChain({ chainId: '0x1' })}
-      >Set Chain to Mainnet</button
-    >
-    <button on:click={() => onboard.setChain({ chainId: '0x4' })}
-      >Set Chain to Rinkeby</button
-    >
-    <button on:click={() => onboard.setChain({ chainId: '0x89' })}
-      >Set Chain to Matic</button
-    >
-  {/if}
+    {#if $wallets$}
+      <button on:click={() => onboard.state.actions.updateBalances()}
+        >Update Wallet Balance</button
+      >
+      <div class="notify-chain-container">
+        <div class="notify-action-container">
+          <button
+            on:click={() =>
+              onboard.state.actions.customNotification({
+                type: 'hint',
+                message: 'This is a custom DApp hint',
+                autoDismiss: 0
+              })}>Send Hint Notification</button
+          >
+          <button
+            on:click={() => {
+              const { update, dismiss } =
+                onboard.state.actions.customNotification({
+                  type: 'pending',
+                  message:
+                    'This is a custom DApp pending notification to use however you want',
+                  autoDismiss: 0
+                })
+              setTimeout(
+                () =>
+                  update({
+                    eventCode: 'dbUpdateSuccess',
+                    message: 'Updated status for custom notification',
+                    type: 'success',
+                    autoDismiss: 0
+                  }),
+                4000
+              )
+            }}>Send Success Notification</button
+          >
+          <button
+            on:click={() =>
+              onboard.state.actions.customNotification({
+                message:
+                  'This is a custom DApp success notification to use however you want',
+                autoDismiss: 0,
+                type: 'pending'
+              })}>Send Pending Notification</button
+          >
+          <button
+            on:click={() =>
+              onboard.state.actions.customNotification({
+                type: 'error',
+                message:
+                  'This is a custom DApp Error notification to use however you want',
+                autoDismiss: 0
+              })}>Send Error Notification</button
+          >
+          <button
+            on:click={() =>
+              onboard.state.actions.customNotification({
+                message:
+                  'This is a custom non-descript DApp notification to use however you want',
+                autoDismiss: 0
+              })}>Send DApp Notification</button
+          >
+        </div>
+        <div class="switch-chain-container">
+          <button on:click={() => onboard.setChain({ chainId: '0x1' })}
+            >Set Chain to Mainnet</button
+          >
+          <button on:click={() => onboard.setChain({ chainId: '0x4' })}
+            >Set Chain to Rinkeby</button
+          >
+          <button on:click={() => onboard.setChain({ chainId: '0x89' })}
+            >Set Chain to Matic</button
+          >
+        </div>
+      </div>
+    {/if}
+  </div>
 
   {#if $wallets$}
     {#each $wallets$ as { icon, label, accounts, chains, provider }}
@@ -263,6 +468,7 @@
 
         {#each accounts as { address, ens, balance }}
           <div
+            class="account-info"
             style="margin-top: 0.25rem; margin-bottom: 0.25rem; padding: 0.25rem; border: 1px solid gray;"
           >
             <div>Address: {address}</div>
@@ -299,6 +505,29 @@
             />
             <button on:click={signTypedMessage(provider, address)}>
               Sign Typed Message
+            </button>
+          </div>
+
+          <div>
+            <input
+              type="text"
+              class="text-input"
+              placeholder="0x..."
+              bind:value={toAddress}
+            />
+            <button on:click={sendTransaction(provider)}>
+              Send Transaction
+            </button>
+          </div>
+          <div>
+            <input
+              type="text"
+              class="text-input"
+              placeholder="0x..."
+              bind:value={toAddress}
+            />
+            <button on:click={sendTransactionWithPreFlight(provider, balance)}>
+              Send with Preflight Notifications
             </button>
           </div>
 
