@@ -1,15 +1,13 @@
-import type {
-  ScanAccountsOptions,
-  Account,
-  Asset,
-  Chain,
-  WalletInit,
-  EIP1193Provider
-} from '@web3-onboard/common'
-
+import type { Chain, Platform, WalletInit } from '@web3-onboard/common'
 import type { StaticJsonRpcProvider } from '@ethersproject/providers'
 import type { ETHAccountPath } from '@shapeshiftoss/hdwallet-core'
 import type { KeepKeyHDWallet } from '@shapeshiftoss/hdwallet-keepkey'
+
+import type {
+  ScanAccountsOptions,
+  Account,
+  Asset
+} from '@web3-onboard/hw-common'
 
 const DEFAULT_PATH = `m/44'/60'/0'/0/0`
 
@@ -38,11 +36,17 @@ const errorMessages = {
 
 type ErrorCode = 'busy' | 'pairing'
 
-function keepkey(): WalletInit {
+function keepkey({ filter }: { filter?: Platform[] } = {}): WalletInit {
   const getIcon = async () => (await import('./icon.js')).default
 
-  return () => {
+  return ({ device }) => {
     let accounts: Account[] | undefined
+
+    const filtered =
+      Array.isArray(filter) &&
+      (filter.includes(device.type) || filter.includes(device.os.name))
+
+    if (filtered) return null
 
     return {
       label: 'KeepKey',
@@ -60,12 +64,18 @@ function keepkey(): WalletInit {
           HDWalletErrorType
         } = await import('@shapeshiftoss/hdwallet-core')
 
-        const {
-          accountSelect,
-          createEIP1193Provider,
-          ProviderRpcError,
-          entryModal
-        } = await import('@web3-onboard/common')
+        const { createEIP1193Provider, ProviderRpcError } = await import(
+          '@web3-onboard/common'
+        )
+
+        const { accountSelect, entryModal } = await import(
+          '@web3-onboard/hw-common'
+        )
+
+        const { bigNumberFieldsToStrings, getHardwareWalletProvider } =
+          await import('@web3-onboard/hw-common')
+
+        const { utils } = await import('ethers')
 
         const { StaticJsonRpcProvider } = await import(
           '@ethersproject/providers'
@@ -164,39 +174,45 @@ function keepkey(): WalletInit {
           asset: Asset
           provider: StaticJsonRpcProvider
         }) => {
-          let index = getAccountIdx(derivationPath)
-          let zeroBalanceAccounts = 0
-          const accounts = []
+          try {
+            let index = getAccountIdx(derivationPath)
+            let zeroBalanceAccounts = 0
+            const accounts = []
 
-          // Iterates until a 0 balance account is found
-          // Then adds 4 more 0 balance accounts to the array
-          while (zeroBalanceAccounts < 5) {
-            const acc = await getAccount({
-              accountIdx: index,
-              provider,
-              asset
-            })
+            // Iterates until a 0 balance account is found
+            // Then adds 4 more 0 balance accounts to the array
+            while (zeroBalanceAccounts < 5) {
+              const acc = await getAccount({
+                accountIdx: index,
+                provider,
+                asset
+              })
 
-            if (
-              acc &&
-              acc.balance &&
-              acc.balance.value &&
-              acc.balance.value.isZero()
-            ) {
-              zeroBalanceAccounts++
-              accounts.push(acc)
-            } else {
-              accounts.push(acc)
-              // Reset the number of 0 balance accounts
-              zeroBalanceAccounts = 0
+              if (
+                acc &&
+                acc.balance &&
+                acc.balance.value &&
+                acc.balance.value.isZero()
+              ) {
+                zeroBalanceAccounts++
+                accounts.push(acc)
+              } else {
+                accounts.push(acc)
+                // Reset the number of 0 balance accounts
+                zeroBalanceAccounts = 0
+              }
+
+              index++
             }
 
-            index++
+            return accounts
+          } catch (error) {
+            throw new Error(
+              (error as { message: { message: string } }).message.message
+            )
           }
-
-          return accounts
         }
-
+        let ethersProvider: StaticJsonRpcProvider
         const scanAccounts = async ({
           derivationPath,
           chainId,
@@ -205,7 +221,7 @@ function keepkey(): WalletInit {
           if (!keepKeyWallet)
             throw new Error('Device must be connected before scanning accounts')
           currentChain = chains.find(({ id }) => id === chainId) || currentChain
-          const provider = new StaticJsonRpcProvider(currentChain.rpcUrl)
+          ethersProvider = new StaticJsonRpcProvider(currentChain.rpcUrl)
 
           // Checks to see if this is a custom derivation path
           // If it is then just return the single account
@@ -214,7 +230,11 @@ function keepkey(): WalletInit {
           ) {
             try {
               const accountIdx = getAccountIdx(derivationPath)
-              const account = await getAccount({ accountIdx, provider, asset })
+              const account = await getAccount({
+                accountIdx,
+                provider: ethersProvider,
+                asset
+              })
 
               return [account]
             } catch (error) {
@@ -222,7 +242,11 @@ function keepkey(): WalletInit {
             }
           }
 
-          return getAllAccounts({ derivationPath, asset, provider })
+          return getAllAccounts({
+            derivationPath,
+            asset,
+            provider: ethersProvider
+          })
         }
 
         const getAccounts = async () => {
@@ -271,27 +295,9 @@ function keepkey(): WalletInit {
           return signature
         }
 
-        const request: EIP1193Provider['request'] = async ({
-          method,
-          params
-        }) => {
-          const response = await fetch(currentChain.rpcUrl, {
-            method: 'POST',
-            body: JSON.stringify({
-              id: '42',
-              method,
-              params
-            })
-          }).then(res => res.json())
-
-          if (response.result) {
-            return response.result
-          } else {
-            throw response.error
-          }
-        }
-
-        const keepKeyProvider = { request }
+        const keepKeyProvider = getHardwareWalletProvider(
+          () => currentChain.rpcUrl
+        )
 
         const provider = createEIP1193Provider(keepKeyProvider, {
           eth_requestAccounts: async () => {
@@ -364,41 +370,75 @@ function keepkey(): WalletInit {
                 'No account selected. Must call eth_requestAccounts first.'
               )
 
+            // Per the code above if accounts is empty or undefined then this line of code won't execute
+            // ∴ account must be defined here which is why it is cast without the 'undefined' type
             const account =
               !transactionObject || !transactionObject.hasOwnProperty('from')
                 ? accounts[0]
-                : accounts.find(
-                    account => account.address === transactionObject.from
-                  )
+                : (accounts.find(
+                    account =>
+                      account.address.toLocaleLowerCase() ===
+                      transactionObject.from.toLocaleLowerCase()
+                  ) as Account)
 
-            const { derivationPath } = account || accounts[0]
+            const { derivationPath, address } = account
             const addressNList = bip32ToAddressNList(derivationPath)
 
+            const signer = ethersProvider.getSigner(address)
+
+            transactionObject.gasLimit =
+              transactionObject.gas || transactionObject.gasLimit
+
+            // 'gas' is an invalid property for the TransactionRequest type
+            delete transactionObject.gas
+
+            transactionObject.gasLimit = undefined
+
+            let populatedTransaction = await signer.populateTransaction(
+              transactionObject
+            )
+
             const {
-              nonce,
-              gasPrice,
-              gas,
-              gasLimit,
               to,
               value,
-              data,
-              maxFeePerGas,
-              maxPriorityFeePerGas
-            } = transactionObject
-
-            const { serialized } = await keepKeyWallet.ethSignTx({
-              addressNList,
-              nonce: nonce || '0x0',
+              nonce,
+              gasLimit,
               gasPrice,
-              gasLimit: gasLimit || gas || '0x5208',
-              to,
-              value: value || '0x0',
-              data: data || '',
               maxFeePerGas,
               maxPriorityFeePerGas,
-              chainId: parseInt(currentChain.id)
-            })
+              data
+            } = bigNumberFieldsToStrings(populatedTransaction)
 
+            const gasData = gasPrice
+              ? {
+                  gasPrice
+                }
+              : {
+                  maxFeePerGas,
+                  maxPriorityFeePerGas
+                }
+
+            const txn = {
+              addressNList,
+              chainId: parseInt(currentChain.id),
+              to: to || '',
+              value: value || '',
+              nonce: utils.hexValue(nonce),
+              gasLimit: gasLimit || '0x0',
+              data: (data || '').toString(),
+              ...gasData
+            }
+
+            let serialized
+            try {
+              ;({ serialized } = await keepKeyWallet.ethSignTx(txn))
+            } catch (error: any) {
+              if (error.message && error.message.message) {
+                throw new Error(error.message.message)
+              } else {
+                throw new Error(error)
+              }
+            }
             return serialized
           },
           eth_sendTransaction: async ({ baseRequest, params }) => {
