@@ -1,12 +1,8 @@
 import { firstValueFrom, Subject, take } from 'rxjs'
 import LoginModal from './view/LoginModal.svelte'
-import { loggedIn$ } from './streams'
-import {
-  SofiaProRegular,
-  SofiaProSemiBold,
-  SofiaProLight
-} from '@web3-onboard/common'
-import type { LoginOptions } from './types'
+import { loggedIn$ } from './streams.js'
+import { SofiaProRegular, SofiaProLight } from '@web3-onboard/common'
+import type { LoginOptions } from './types.js'
 
 // eslint-disable-next-line max-len
 const loginModal = async (options: LoginOptions): Promise<boolean> => {
@@ -49,7 +45,6 @@ const mountLoginModal = (
 
   styleEl.innerHTML = `
     ${SofiaProRegular}
-    ${SofiaProSemiBold}
     ${SofiaProLight}
   `
   document.body.appendChild(styleEl)
