@@ -1,12 +1,12 @@
 export type LoginOptions = {
-  walletName: string,
-  brandingHTMLString: string,
+  walletName: string
+  brandingHTMLString: string
   emailLoginFunction: EmailLoginFunction
 }
 
 export type MagicInitOptions = {
-  apiKey: string,
-  userEmail?: string,
+  apiKey: string
+  userEmail?: string
 }
 
 export type EmailLoginFunction = (emailAddress: string) => Promise<boolean>
